JPA repository, dependency injection, database seeding

+++ src/main/java/com/educandoweb/course2/entities/User.java
@@ -17,7 +17,7 @@ public class User  implements Serializable{
     
     @Id //Declarando chave primaria
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
     private String name;
     private String email;
     private String phone;
@@ -28,7 +28,7 @@ public class User  implements Serializable{
     }
 
     //COSNTRUTOR SOBRECARREGADO
-    public User(long id, String name, String email, String phone, String password) {
+    public User(Integer id, String name, String email, String phone, String password) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -64,11 +64,11 @@ public class User  implements Serializable{
 
     
     //GET AND SET
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
